safari-534.53-branch/LayoutTests/fast/canvas/canvas-invalid-strokestyle-expected.txt

Series of tests to ensure correct behaviour on an invalid strokeStyle()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[4] is 0
PASS imgdata[5] is 255
PASS imgdata[6] is 0
PASS successfullyParsed is true

TEST COMPLETE
